requirements.txt
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1733247046149/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/anyio_1736174388474/work
bcrypt==4.3.0
cassandra-driver==3.29.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
cffi==1.17.1
click @ file:///D:/bld/click_1734858834134/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
cryptography==44.0.2
dnspython @ file:///home/conda/feedstock_root/build_artifacts/dnspython_1733256735222/work
ecdsa==0.19.1
email_validator @ file:///home/conda/feedstock_root/build_artifacts/email-validator-meta_1733300719943/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
fastapi @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_fastapi_1740930998/work
fastapi-cli @ file:///home/conda/feedstock_root/build_artifacts/fastapi-cli_1734302308128/work
geomet @ file:///home/conda/feedstock_root/build_artifacts/geomet_1601766976124/work
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1733327467879/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_httpcore_1731707562/work
httptools @ file:///D:/bld/httptools_1732707690819/work
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1733663348460/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1734823942230/work
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1733250460757/work
MarkupSafe @ file:///D:/bld/markupsafe_1733219614059/work
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
numpy @ file:///D:/bld/numpy_1739425617249/work/dist/numpy-2.2.3-cp311-cp311-win_amd64.whl#sha256=16d8179eb346ade4b011e196728dfd85a6ee21445113538d9be6c1363856be27
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic @ file:///home/conda/feedstock_root/build_artifacts/pydantic_1737761369378/work
pydantic_core @ file:///D:/bld/pydantic-core_1734571360609/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/python-dotenv_1733243180666/work
python-jose==3.4.0
python-multipart @ file:///home/conda/feedstock_root/build_artifacts/python-multipart_1734420773152/work
PyYAML @ file:///D:/bld/pyyaml_1737454660454/work
rich @ file:///home/conda/feedstock_root/build_artifacts/rich_1733342254348/work/dist
rich-toolkit @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich-toolkit_1733750834/work
rsa==4.9
shellingham @ file:///home/conda/feedstock_root/build_artifacts/shellingham_1733300899265/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1733244044561/work
starlette @ file:///home/conda/feedstock_root/build_artifacts/starlette_1740260770515/work
typer==0.15.2
typer-slim==0.15.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
uvicorn @ file:///D:/bld/uvicorn_1734292951833/work
watchfiles @ file:///D:/bld/watchfiles_1736550434446/work
websockets @ file:///D:/bld/websockets_1739780540187/work
